+++ .env
@@ -36,5 +36,5 @@ MESSENGER_TRANSPORT_DSN=doctrine://default?auto_setup=0
 ###< symfony/messenger ###
 
 ###> symfony/mailer ###
-MAILER_DSN=null://null
+MAILER_DSN=gmail+smtp://USERNAME:APP-PASSWORD@default
 ###< symfony/mailer ###
