+++ src/Entity/Category.php
@@ -16,7 +16,7 @@ class Category
     #[ORM\Column]
     private ?int $id = null;
 
-    #[ORM\Column(length: 255)]
+    #[ORM\Column(length: 255, unique: true )]
     private ?string $name = null;
 
     #[ORM\Column(type: Types::GUID)]
