+++ src/Entity/Trick.php
@@ -28,7 +28,7 @@ class Trick
     #[ORM\Column(nullable:true)]
     private ?\DateTimeImmutable $updatedAt = null;
 
-    #[ORM\Column(length: 255)]
+    #[ORM\Column(length: 255, unique: true)]
     private ?string $slug = null;
 
     #[ORM\ManyToOne(inversedBy: 'tricks')]
