+++ symfony.lock
@@ -288,9 +288,6 @@
             "./config/packages/messenger.yaml"
         ]
     },
-    "symfonycasts/verify-email-bundle": {
-        "version": "v1.13.0"
-    },
     "twig/extra-bundle": {
         "version": "v3.6.0"
     }
